:art: refactor(directory): 디렉토리 구조 정리
- racingcar 디렉토리 : Application, Game
  - data 디렉토리 : GameData
    - attempt 디렉토리 : (attempt 관련 클래스들), Move
    - car 디렉토리 : (car 관련 클래스들)
  - execution 디렉토리 : (input, output 관련 클래스들)
