+++ src/main/java/racingcar/Application.java
@@ -1,4 +1,4 @@
-package racingcar.game;
+package racingcar;
 
 public class Application {
     public static void main(String[] args) {
