+++ src/main/java/racingcar/data/attempt/AttemptNumber.java
@@ -1,4 +1,4 @@
-package racingcar.data;
+package racingcar.data.attempt;
 
 public class AttemptNumber {
     private final int number;
