+++ src/main/java/racingcar/data/attempt/Move.java
@@ -1,4 +1,4 @@
-package racingcar.data;
+package racingcar.data.attempt;
 
 public enum Move {
     STOP(0),
