+++ src/main/java/racingcar/data/attempt/MoveAttempts.java
@@ -1,4 +1,4 @@
-package racingcar.data;
+package racingcar.data.attempt;
 
 import java.util.ArrayList;
 import java.util.List;
