+++ src/main/java/racingcar/data/car/CarPosition.java
@@ -1,4 +1,4 @@
-package racingcar.data;
+package racingcar.data.car;
 
 public class CarPosition {
     private int pos;
