+++ src/main/java/racingcar/data/car/RacingCars.java
@@ -1,4 +1,4 @@
-package racingcar.data;
+package racingcar.data.car;
 
 import java.util.ArrayList;
 import java.util.List;
